feat:initialize solution structure for CF231C

+++ C.ToAddOrNotToAdd/CF231C/CF231CCToAddOrNotToAdd.java
@@ -23,7 +23,7 @@ public class CF231CCToAddOrNotToAdd {
             String ValuesLine = bufferedReader.readLine();
             StringTokenizer valueTokenizer = new StringTokenizer(ValuesLine, " ");
             int[] nums = new int[n];
-            int bestValue = nums[0];
+             long bestValue = nums[0];
             for (int j = 0; valueTokenizer.hasMoreTokens(); j++) {
                 nums[j] = Integer.parseInt(valueTokenizer.nextToken());
             }
